Improve user experience of HostedCluster configuration API
Currently, global configuration of HostedClusters needs to be specified
as an arbitrary list of RawExtensions. This forces users to have to
properly specify kind and apiVersion on basic configuration resources.
In addition to that, redundant information needs to be provided as a
list of secret and configmap references.

This change simplifies hosted cluster configuration by including
specific fields for each configuration resource's spec. The user has
now explicit fields where they can specify configuration for their
cluster.